Rename 'fixed_size_shards' dir to 'shards'
